Separa queries por módulos de consulta
